requirements.txt
typeform
sqlite3
anthropic
pypdf2
python-docx